pyproject.toml
[tool.poetry]
name = "nikoli-puzzle-solver"
version = "0.1.0"
description = ""
authors = ["Kevin Y. Chen"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.2"
grilops = "^0.9.1"
pytest = "^7.2.0"
"ruamel.yaml" = "^0.17.21"
z3-solver = "^4.11.2"
black = "^22.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
